// src/lib/types/Character.ts
export interface Character {
	char_name?: string;
	char_birthday?: string;
	location?: string;
	ethnicity?: string;
	hair_style?: string;
	hair_color?: string;
	eye_color?: string;
	breast_size?: string;
	body_type?: string;
	butt_size?: string;
	age?: string;
	gender?: string;
	occupation?: string;
	hobby_1?: string;
	hobby_2?: string;
	personality_name?: string;
	personality_description?: string;
}